PC-1912: Add function to manual reminder email job
for visibility to DESNZ, includes a notice after pressing the button for success
