+++ WhlgPortalWebsite/Models/ServiceManagerHomepageViewModel.cs
@@ -8,7 +8,8 @@ public class ServiceManagerHomepageViewModel
 {
     public string SearchEmailAddress { get; }
     public IEnumerable<AuthorityUserListing> UserList { get; }
-    public bool ShouldShowManualJobRunner { get; set; }
+    public bool ShowJobSuccess { get; set; }
+    public bool ShowManualJobRunner { get; set; }
 
     public ServiceManagerHomepageViewModel(IEnumerable<User> users)
     {
